+++ Sources/VComponents/Components/Indicators (Definite)/Page Indicator (Compact)/VCompactPageIndicator.swift
@@ -209,11 +209,11 @@ public struct VCompactPageIndicator<Content>: View where Content: View {
             switch dotContent {
             case .empty:
                 ZStack(content: {
-                    Capsule()
+                    RoundedRectangle(cornerRadius: uiModel.dotCornerRadii.value(for: internalState))
                         .foregroundColor(uiModel.dotColors.value(for: internalState))
                     
                     if uiModel.dotBorderWidths.value(for: internalState) > 0 {
-                        Capsule()
+                        RoundedRectangle(cornerRadius: uiModel.dotCornerRadii.value(for: internalState))
                             .strokeBorder(lineWidth: uiModel.dotBorderWidths.value(for: internalState))
                             .foregroundColor(uiModel.dotBorderColors.value(for: internalState))
                     }